Combining GlobalExceptionHandlers
